PYTHON-6023 Consolidate all non-IO code in AsyncCursor/Cursor into _AgnosticCursor
