+++ pymongo/typings.py
@@ -31,12 +31,14 @@ if TYPE_CHECKING:
     from pymongo.asynchronous.bulk import _AsyncBulk
     from pymongo.asynchronous.client_bulk import _AsyncClientBulk
     from pymongo.asynchronous.client_session import AsyncClientSession
+    from pymongo.asynchronous.collection import AsyncCollection
     from pymongo.asynchronous.mongo_client import AsyncMongoClient
     from pymongo.asynchronous.pool import AsyncConnection
     from pymongo.collation import Collation
     from pymongo.synchronous.bulk import _Bulk
     from pymongo.synchronous.client_bulk import _ClientBulk
     from pymongo.synchronous.client_session import ClientSession
+    from pymongo.synchronous.collection import Collection
     from pymongo.synchronous.mongo_client import MongoClient
     from pymongo.synchronous.pool import Connection
 
@@ -51,6 +53,7 @@ _T = TypeVar("_T")
 
 # Type hinting types for compatibility between async and sync classes
 _AgnosticMongoClient = Union["AsyncMongoClient", "MongoClient"]  # type: ignore[type-arg]
+_AgnosticCollection = Union["AsyncCollection[_DocumentType]", "Collection[_DocumentType]"]
 _AgnosticConnection = Union["AsyncConnection", "Connection"]
 _AgnosticClientSession = Union["AsyncClientSession", "ClientSession"]
 _AgnosticBulk = Union["_AsyncBulk", "_Bulk"]
